heartbeat/requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
apscheduler>=3.10.0
python-dateutil>=2.8.0
pyyaml>=6.0
